Desktop/botbirzha/requirements.txt
aiofiles==22.1.0
pyTelegramBotAPI==4.7.0
PyMySQL==1.0.2
aiogram-broadcaster==0.0.7
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1